# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "数智化管理系统",
    "description": "数智化管理系统 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/invoice/ocr/upload": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "上传 + 单张识别",
        "operationId": "upload_api_v1_invoice_ocr_upload_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_invoice_ocr_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/list": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "发票识别记录列表",
        "operationId": "list_invoices_api_v1_invoice_ocr_list_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/detail": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "发票详情",
        "operationId": "detail_api_v1_invoice_ocr_detail_post",
        "parameters": [
          {
            "name": "invoiceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoiceid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/update": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "编辑/核验发票",
        "operationId": "update_api_v1_invoice_ocr_update_post",
        "parameters": [
          {
            "name": "invoiceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoiceid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Body_update_api_v1_invoice_ocr_update_post"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/submit": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "提交入账",
        "operationId": "submit_api_v1_invoice_ocr_submit_post",
        "parameters": [
          {
            "name": "invoiceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoiceid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/recheck": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "重新识别",
        "operationId": "recheck_api_v1_invoice_ocr_recheck_post",
        "parameters": [
          {
            "name": "invoiceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoiceid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/batch/upload": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "批量上传",
        "operationId": "batch_upload_api_v1_invoice_ocr_batch_upload_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_batch_upload_api_v1_invoice_ocr_batch_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/batch/status": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "批量任务状态（兼容旧版，建议用 SSE）",
        "operationId": "batch_status_api_v1_invoice_ocr_batch_status_post",
        "parameters": [
          {
            "name": "batchId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "批量任务 ID（code 字符串）",
              "title": "Batchid"
            },
            "description": "批量任务 ID（code 字符串）"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/batch/submit": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "批量提交入账",
        "operationId": "batch_submit_api_v1_invoice_ocr_batch_submit_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/batch/retry": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "批量重试失败项",
        "operationId": "batch_retry_api_v1_invoice_ocr_batch_retry_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRetryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/advanced-search": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "高级筛选",
        "operationId": "advanced_search_api_v1_invoice_ocr_advanced_search_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/batch-action": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "批量操作",
        "operationId": "batch_action_api_v1_invoice_ocr_batch_action_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/saved-view": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "保存为常用视图",
        "operationId": "save_view_api_v1_invoice_ocr_saved_view_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedViewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoice/ocr/views": {
      "post": {
        "tags": [
          "发票识别"
        ],
        "summary": "已保存的视图",
        "operationId": "list_views_api_v1_invoice_ocr_views_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchActionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "invoiceIds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Invoiceids"
          },
          "params": {
            "type": "object",
            "title": "Params",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "action",
          "invoiceIds"
        ],
        "title": "BatchActionRequest"
      },
      "BatchRetryRequest": {
        "properties": {
          "batchId": {
            "type": "string",
            "title": "Batchid"
          },
          "fileIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fileids"
          }
        },
        "type": "object",
        "required": [
          "batchId",
          "fileIds"
        ],
        "title": "BatchRetryRequest"
      },
      "BatchSubmitRequest": {
        "properties": {
          "invoiceIds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Invoiceids"
          }
        },
        "type": "object",
        "required": [
          "invoiceIds"
        ],
        "title": "BatchSubmitRequest"
      },
      "Body_batch_upload_api_v1_invoice_ocr_batch_upload_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          },
          "templateId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templateid"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_batch_upload_api_v1_invoice_ocr_batch_upload_post"
      },
...
      "Body_update_api_v1_invoice_ocr_update_post": {
        "properties": {
          "fields": {
            "type": "object",
            "title": "Fields",
            "default": {}
          },
          "expenseInfo": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expenseinfo"
          }
        },
        "type": "object",
        "title": "Body_update_api_v1_invoice_ocr_update_post"
      },
...
      "Body_upload_api_v1_invoice_ocr_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_invoice_ocr_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvoiceListRequest": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize",
            "default": 20
          },
          "keyword": {
            "type": "string",
            "title": "Keyword",
            "default": ""
          },
          "filters": {
            "type": "object",
            "title": "Filters",
            "default": {}
          }
        },
        "type": "object",
        "title": "InvoiceListRequest"
      },
...
      "SavedViewRequest": {
        "properties": {
          "viewName": {
            "type": "string",
            "title": "Viewname"
          },
          "filters": {
            "type": "object",
            "title": "Filters"
          },
          "isShared": {
            "type": "boolean",
            "title": "Isshared",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "viewName",
          "filters"
        ],
        "title": "SavedViewRequest"
      },
...
    }
  }
}